requirements.txt
flask
flask-mail
mysql-connector-python
python-dotenv